add price count

+++ app/Http/Controllers/BasketController.php
@@ -33,7 +33,13 @@ class BasketController extends Controller
             $order = Order::find($orderId); // Находим существующий заказ по идентификатору
         }
 
-        $order->products()->attach($productId); // Прикрепляем продукт к заказу
+        if ($order->products->contains($productId)) {
+            $pivotRow = $order->products()->where('product_id', $productId)->first()->pivot;
+            $pivotRow->count++;
+            $pivotRow->update();
+        } else {
+            $order->products()->attach($productId);
+        }
 
         return redirect()->route('basket');
 
@@ -46,7 +52,17 @@ class BasketController extends Controller
             return redirect()->route('basket');
         }
         $order = Order::find($orderId);
-        $order->products()->detach($productId);
+
+        if ($order->products->contains($productId)) {
+            $pivotRow = $order->products()->where('product_id', $productId)->first()->pivot;
+            if ($pivotRow->count < 2) {
+                $order->products()->detach($productId);
+            } else {
+                $pivotRow->count--;
+                $pivotRow->update();
+            }
+        }
+
 
         return redirect()->route('basket');
